Moved IField and utils to utils dir. Reorganised React markup so it's easier to follow. Reordered cert fields properly. Now catering dynamically for different types of radio in the vessel details. Added padding-left to the copy to clipboard button
