+++ backoffice/src/certificate.scss
@@ -26,6 +26,11 @@
   transform-origin: 60%;
 }
 
+.subItem {
+  padding-top: 1%;
+  padding-bottom: 1%;
+}
+
 .title {
   font-weight: bold;
 }
