+++ backoffice/src/fixtures/emergencyContacts.fixture.ts
@@ -1,5 +1,5 @@
 import { IEmergencyContact } from "../entities/IEmergencyContact";
-import { deepFreeze } from "../utils";
+import { deepFreeze } from "../utils/utils";
 
 export const emergencyContactsFixture = deepFreeze<IEmergencyContact[]>([
   {
